feat: admin mode toggle (#133) - Octavio H.
* Add: fixed some health data bugs and updates to calories in food diary sync to health data

* update: frontend works with new health data fixes, MealViewModel works with endpoints, and Today's Overview prompt update

* feat: implemented admin sessions and challenge  insert/update/delete endpoints only accessible to admin

* add: DataClient methods to handle new endpoints and updating to UserData model in frontend

* add: upsert endpoint for challenges, unit and integration testing for challenges collection

* add: admin verification with password prompt and challenge editor

* update: install mongoDB dependencies in dotnet-ci to avoid errors

* update: fix test compiling issues caused by UserData updates

* add: toggle button for admin mode features

+++ Frontend/composeApp/src/commonMain/kotlin/com/teamnotfound/airise/community/challenges/ChallengesScreen.kt
@@ -42,6 +42,7 @@ import com.teamnotfound.airise.util.White
 import com.teamnotfound.airise.AppScreen
 import com.teamnotfound.airise.auth.admin.AdminVerifyViewModel
 import com.teamnotfound.airise.auth.admin.AdminVerifyDialog
+import com.teamnotfound.airise.auth.admin.AdminVerifyUiEvent
 import com.teamnotfound.airise.auth.admin.AdminVerifyUiState
 import com.teamnotfound.airise.community.challenges.challengeEditor.ChallengeEditorDialog
 import com.teamnotfound.airise.community.challenges.challengeEditor.ChallengeEditorUiEvent
@@ -115,27 +116,41 @@ fun ChallengesScreen(
                 horizontalArrangement = Arrangement.spacedBy(12.dp),
                 verticalAlignment = Alignment.CenterVertically
             ) {
+
                 if (adminState.isAdmin) {
-                    // Add Challenge FAB (Admin only)
-                    ExtendedFloatingActionButton(
-                        text = { Text("Add Challenge") },
+                    ExtendedFloatingActionButton(text = { Text("Admin Mode") },
                         onClick = {
                             if (adminState.isVerified) {
-                                editorViewModel.onEvent(
-                                    ChallengeEditorUiEvent.OpenEditor(
-                                        ChallengeUI()
-                                    )
-                                )
+                                adminVerifyViewModel.onEvent(AdminVerifyUiEvent.ToggleAdminModeActive)
                             } else {
                                 // Not verified, show password dialog
                                 adminVerifyViewModel.showPasswordPrompt()
                             }
                         },
                         backgroundColor = Orange,
-                        contentColor = BgBlack
+                        contentColor = BgBlack,
                     )
+                    if (adminState.isAdminModeActive){
+                        // Add Challenge FAB (Admin only)
+                        ExtendedFloatingActionButton(
+                            text = { Text("Add Challenge") },
+                            onClick = {
+                                if (adminState.isAdminModeActive && adminState.isVerified) {
+                                    editorViewModel.onEvent(
+                                        ChallengeEditorUiEvent.OpenEditor(
+                                            ChallengeUI()
+                                        )
+                                    )
+                                } else {
+                                    // Not verified, show password dialog
+                                    adminVerifyViewModel.showPasswordPrompt()
+                                }
+                            },
+                            backgroundColor = Orange,
+                            contentColor = BgBlack
+                        )
+                    }
                 }
-
 
                 FloatingActionButton(
                     onClick = {
@@ -295,7 +310,7 @@ private fun ChallengeCard(
                         modifier = Modifier.fillMaxSize()
                     )
 
-                    if(adminState.isAdmin){
+                    if(adminState.isAdminModeActive && adminState.isAdmin){
                         Box(
                             modifier = Modifier
                                 .matchParentSize()
@@ -309,7 +324,7 @@ private fun ChallengeCard(
                         ) {
                             IconButton(
                                 onClick = {
-                                    if (adminState.isVerified) {
+                                    if (adminState.isAdminModeActive && adminState.isVerified) {
                                         editorViewModel.delete(challengeUI.id)
                                     } else {
                                         // Not verified, show password dialog
@@ -331,7 +346,7 @@ private fun ChallengeCard(
                             }
                             IconButton(
                                 onClick = {
-                                    if (adminState.isVerified) {
+                                    if (adminState.isAdminModeActive && adminState.isVerified) {
                                         editorViewModel.onEvent(
                                             ChallengeEditorUiEvent.OpenEditor(
                                                 challengeUI
